requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
playwright>=1.42.0
pydantic==2.4.2
pydantic-settings==2.1.0
sqlalchemy==2.0.27
aiosqlite>=0.19.0
python-dotenv==1.0.1
aiohttp>=3.9.1
async-timeout>=4.0.3
typing-extensions>=4.8.0
